Rename best/strings -> best/text
